feat: Add AppError class and error codes
This commit adds the `AppError` class and error codes to the `common` module. The `AppError` class extends the `Error` class and includes a code and detail message. The error codes are defined using the `arrayToEnum` function. This addition will provide a standardized way to handle and communicate errors in the application.

Refactor: Update auth.grpc-controller

This commit updates the `auth.grpc-controller` file in the `infa/controllers` module. It adds error handling for invalid tokens by throwing a `RpcException` with the code "invalid_token" and a corresponding message. It also caches the token as "invalid_token" for 1 minute to prevent unnecessary calls to the `userInfo` service. This improvement enhances the security and performance of the authentication process.

Refactor: Rename checks method in health.controller

This commit renames the `check` method in the `health.controller` file in the `infa/controllers` module to `checks`. This change improves the clarity and consistency of the codebase.

Refactor: Add HttpExceptionFilter and RpcExceptionFilter

This commit adds the `HttpExceptionFilter` and `RpcExceptionFilter` classes to the `filters` module. The `HttpExceptionFilter` handles all exceptions thrown in the HTTP context and returns a standardized error response. The `RpcExceptionFilter` handles `RpcException` specifically. These filters improve the error handling and response consistency in the application.

Refactor: Update main.ts

This commit updates the `main.ts` file in the root directory. It replaces the `Error` class import with `ForbiddenException` from `@nestjs/common` to handle CORS errors. This change ensures that the appropriate exception is thrown when a request is not allowed by CORS.

Refactor: Add Authenticated guard

This commit adds the `Authenticated` guard to the `guards` module. The guard checks for the presence of an authorization token in the request headers and verifies it using the `verifyToken` method from the `AuthService`. If the token is valid, the guard sets the user information in the request object. This guard enhances the security of the application by ensuring that only authenticated users can access protected routes.