chore: Remove unused Kong configuration file
Remove the unused Kong configuration file at apps/api-gateway/kong/config/kong.yaml. This file was no longer needed and has been deleted.

Refactor the storage-service app.config.ts file

In the storage-service app, refactor the app.config.ts file. Import the zod library and register the ConfigType and AppConfig types. This improves the type safety and readability of the configuration file.

Update dependencies in auth-service package.json

In the auth-service app, update the dependencies in the package.json file. Add the express package with version 4.21.0. This is required for some new functionality in the app.

Add a new method in auth-service auth.grpc-controller.ts

In the auth-service app, add a new method called verifyTokenv2 in the auth.grpc-controller.ts file. This method will be used for a specific use case related to token verification.

Refactor the PrismaModule in storage-service

In the storage-service app, refactor the PrismaModule in the prisma.module.ts file. Instead of providing the PrismaClient directly, use a provider object with a useValue property. This allows for more flexibility in configuring the PrismaClient.

Update the health.controller.ts in auth-service

In the auth-service app, update the health.controller.ts file. Add the ApiBearerAuth decorator to the checks method to require authentication for accessing the health endpoint. Also, log the request headers for debugging purposes.

Add new functionality to user.rest-controller.ts in auth-service

In the auth-service app, add new functionality to the user.rest-controller.ts file. Update the list method to retrieve the authenticated user ID from the request headers and use it to fetch the user data. Return an empty response for now.

Refactor the main.ts file in storage-service

In the storage-service app, refactor the main.ts file. Remove unused imports and update the buildMicroservice and buildRmq functions to use the ConfigService from the app module.

Update the HTTPLogger middleware in auth-service

In the auth-service app, update the HTTPLogger middleware in the http-logger.middleware.ts file. Log the client IP, request details, and response details for each HTTP request.

Refactor the config.module.ts file in storage-service

In the storage-service app, refactor the config.module.ts file. Update the ConfigService to extend the NestConfigService and add a validate function to validate and override the environment variables with values from Vault.