chore: Update folder controllers and exports
This commit refactors the folder controllers for gRPC and RESTful APIs. It also updates the exports in the common and storage-client modules. The changes include adding a new folder controller for version 2 (v2) and updating the index file to include both v1 and v2 commands. Additionally, the code removes unused API configuration files and updates dependencies. The configuration.ts file is also updated to include grpcConfigSchema.